New rake task for creating image thumbnails and other rake task modifications.
Modified "new_post" rake task to create a folder for images for that
post on-request in images/blog.  Added a new rake task "img_thumbs"
that looks for images in folders in images/blog, and then creates
thumbnails of those images in a "thumbnails" folder nested inside the
original image's folder, but only if a thumbnail doesn't already exist.
 Added execution of "img_thumbs" to the default "generate" task and
cleaned up some of the output.  Also added RMagick dependency (for
image thumbnail creation).